MFC: r346717: Make bhyve SMBIOS table topology aware
When the CPU Topology was added to bhyve in r332298 the SMBIOS table was
missed, this table passes topology information to the system and was still
using the old concept of each vCPU is a socket with 1 core and 1 thread.
This code did not even try to use the old sysctl information to adjust
this data.

Correct that by building a proper SMBios table, mapping the > 254 cases to
0 per the SMBios 2.6 specification that is claimed by the structure.

Approved by:		re (kib)


git-svn-id: https://svn.freebsd.org/base/stable/11@348269 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f