MFC r348127: bectl(8): add description for create subcommand
In commit r345845, a portion of documentation for the create subcommand was
removed. Specifically, for creating a snapshot of an existing boot
environment. bectl even has a test-case for this functionality.

Removing the sub-command description was discussed in PR 235850.

This patch brings back the second "create" description that was originally
in place. Albeit, with a few wording/clarifying changes.

Approved by:	re (marius)


git-svn-id: https://svn.freebsd.org/base/stable/11@348305 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f