MFC r347642:
The driver list prints "(null)" for the NDIS driver when -h (help) or
an unknown switch is passed outputting the command usage. This is
because the NDIS driver is uninitialized when usage help is printed.
To resolve this we initialize the driver prior to the possibility of
printing the usage help message.

Obtained from:	The wpa_supplicant port


git-svn-id: https://svn.freebsd.org/base/stable/12@348321 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f