Update ena-com for the ENAv2 version
The new ena-com is required for the ENAv2 release of the driver.

Genmask macro was fixed in the platform file.

The GENMASK macro is intended to work on 32-bit values. Before this
patch the, calling GENMASK(31, 0) would cause 32-bit value to be
shifted by 32. This could lead to unexpected behavior, as shifting by
the number greater or equal to word length is undefined operation.

Submitted by:  Michal Krawczyk <mk@semihalf.com>
Obtained from: Semihalf
Sponsored by:  Amazon, Inc.


git-svn-id: https://svn.freebsd.org/base/vendor-sys/ena-com/dist@348383 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f