MFC r348509: jail_getid(3): add special-case immediate return for jid 0
As depicted in the comment: jid 0 always exists, but the lookup will fail as
it does not appear in the kernel's alljails list being a special jail. Some
callers will expect/rely on this, and we have no reason to lie because it
does always exist.

Approved by:	re (gjb)


git-svn-id: https://svn.freebsd.org/base/stable/11@348546 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f