ixl(4)/ixlv(4): Update Intel XL710 PF and VF drivers to ixl-1.11.9 and ixlv-1.5.8
Update the legacy (non-iflib) drivers in stable/11 with recent changes from the
Intel out-of-tree version.

Major changes:

- Support for new BASE-T device with additional link speeds (2.5G and 5G) and EEE
- Additional I2C access methods backported from ixl-iflib
- FW LLDP Agent control with sysctl added for X722 devices (this already
  existed for 710 devices)
- MAC/VLAN filters handling has been refactored
- Building and loading if_ixlv as a KLD has been fixed

This is not a MFC since the driver in 12/13 uses iflib, and the decision was
made to not use it in FreeBSD 11 releases.

Submitted by:	Krzysztof Galazka <krzysztof.galazka@intel.com>
Sponsored by:	Intel Corporation
Differential Revision:	https://reviews.freebsd.org/D20290


git-svn-id: https://svn.freebsd.org/base/stable/11@349163 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f