drm2/intel_iic: stop using iicbus_set_nostop
The desired mode of transmitting messages is implemented by subclassing
iicbb driver and overriding its iicbus_transfer method with an almost
identical copy that issues the stop condition only at the very end.

iicbus_set_nostop is very broken and is set to be removed from the KPI.

This is a direct commit as in head the drm drivers have been moved out
of the tree.
The same change has been committed to FreeBSDDesktop/drm-legacy.


git-svn-id: https://svn.freebsd.org/base/stable/12@349224 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f