MFC r348740-r348741, r348995
r348740:
Don't refer to the cpu variable in a KASSERT before initializing it.

r348741:
For armv6 and armv7, build hwpmc_armv7.c as well as the base hwpmc_arm.c.

Submitted by:	Arnaud YSMAL <arnaud.ysmal@stormshield.eu>

r348995:
Don't attempt to include hwpmc support for armv6, we're missing some of the
necessary support functions in cpu-v6.h, and it may be that the only armv6
platform we support (RPi, the bcm2835 SOC) is incapable of supporting hwpmc.

Reported by:	dim@


git-svn-id: https://svn.freebsd.org/base/stable/12@349315 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f