MFC r340130, r340134:
r340130: (by oshogbo)

libcapsicum: Introduce caph_{rights,ioctls,fcntls}_limit

The idea behind those functions is not to force consumers to remember that there
is a need to check errno on failure. We already have a caph_enter(3) function
which does the same for cap_enter(2).

r340134: (by oshogbo)

Fix a recusive call introduce in the r340130.


git-svn-id: https://svn.freebsd.org/base/stable/12@349425 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f