MFC r349376: Fix strsep_quote() on strings without quotes.
For strings without quotes and escapes dstptr and srcptr are equal, so
zeroing *dstptr before checking *srcptr is not a good idea.  In practice
it means that in -maproot=65534:65533 everything after the colon is lost.

The problem was there since r293305, but before r346976 it was covered by
improper strsep_quote() usage.

PR:	238725


git-svn-id: https://svn.freebsd.org/base/stable/12@349456 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f