More fully implement the state machine.
When a command is finished running, we must transition it from INQUEUE
to busy state. We were failing to do that, so we hit a panic when the
commands were freed. This only affects mpr, mps already did simmilar
things. Now both the polling and interrupt paths properly set BUSY as
appropriate.


git-svn-id: https://svn.freebsd.org/base/head@349909 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f