MFC r348987, r348989:
Resolve IPv6 checksum errors with stateful inspection. According to
PR/203585 this appears to have been broken by r235959, which predates
the ipfilter 5.1.2 import into FreeBSD.

The IPv6 checksum calculation is incorrect. To resolve this we call
in6_cksum() to do the the heavy lifting for us, through a new function
ipf_pcksum6(). Should we need to revisit this area again, a DTrace probe
is added to aid with future debugging.

Plus whitespace adjustments (r348989).

PR:		203275, 203585
Differential Revision:	https://reviews.freebsd.org/D20583


git-svn-id: https://svn.freebsd.org/base/stable/12@349927 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f