MFC r349428,r349433: owc_gpiobus: clean / fix up the driver module things
"fdt" is removed from the driver module name as the driver does not
require FDT and can work very well on hints based systems.

A module dependency is added for gpiobus.  Without that owc cannot
resolve symbols in gpiobus if both are loaded as kernel modules.

Finally, a driver module module version is added.


git-svn-id: https://svn.freebsd.org/base/stable/11@350042 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f